+++ framework/sac_rlhf.py
@@ -4,17 +4,13 @@ import random
 import time
 from dataclasses import dataclass
 import logging
+from typing import Any
 
 import gymnasium as gym
-from gymnasium.experimental.wrappers.rendering import RecordVideoV0
 import numpy as np
 import torch
-import torch.nn as nn
-import torch.nn.functional as F
 import torch.optim as optim
 import tyro
-import pickle
-import gzip
 
 from performance_metrics import PerformanceMetrics
 from video_recorder import VideoRecorder
@@ -25,6 +21,15 @@ from reward_net import RewardNet, train_reward
 from torch.utils.tensorboard import SummaryWriter
 from teacher import Teacher
 from sampling import uniform_sampling, disagreement_sampling, entropy_sampling
+from actor import Actor, select_actions, update_actor, update_target_networks
+from env import (
+    make_env,
+    save_replay_buffer,
+    save_model_all,
+    load_replay_buffer,
+    load_model_all,
+)
+from critic import SoftQNetwork, train_q_network
 
 
 @dataclass
@@ -45,7 +50,7 @@ class Args:
     """the entity (team) of wandb's project"""
     capture_video: bool = False
     """whether to capture videos of the agent performances (check out `videos` folder)"""
-    num_envs: int = 12
+    num_envs: int = 1
     """the number of parallel environments to accelerate training. 
     Set this to the number of available CPU threads for best performance."""
     log_file: bool = True
@@ -58,7 +63,7 @@ class Args:
     """the environment id of the task"""
     total_timesteps: int = 1000000
     """total timesteps of the experiments"""
-    buffer_size: int = int(1e6)
+    buffer_size: int = int(total_timesteps)
     """the replay memory buffer size"""
     gamma: float = 0.99
     """the discount factor gamma"""
@@ -142,226 +147,7 @@ class Args:
     """path to model"""
 
 
-def make_env(env_id, seed, idx, capture_video, run_name):
-    def thunk():
-        if capture_video and idx == 0:
-            env = gym.make(env_id, render_mode="rgb_array")
-            # You've to use experimental wrappers to record video to avoid black screen issue:
-            # https://github.com/Farama-Foundation/Gymnasium/issues/455#issuecomment-1517900688
-            env = RecordVideoV0(env, f"videos/{run_name}")
-        else:
-            env = gym.make(env_id)
-        env = gym.wrappers.RecordEpisodeStatistics(env)
-        env.action_space.seed(seed)
-        return env
-
-    return thunk
-
-
-def select_actions(obs, actor, device, step, learning_start, envs):
-    if step < learning_start:
-        return np.array(
-            [envs.single_action_space.sample() for _ in range(envs.num_envs)]
-        )
-    else:
-        actions, _, _ = actor.get_action(torch.Tensor(obs).to(device))
-        return actions.detach().cpu().numpy()
-
-
-def train_q_network(data, qf1, qf2, qf1_target, qf2_target, alpha, gamma, q_optimizer):
-    with torch.no_grad():
-        real_rewards = data.rewards
-        next_state_actions, next_state_log_pi, _ = actor.get_action(
-            data.next_observations
-        )
-        qf1_next_target = qf1_target(data.next_observations, next_state_actions)
-        qf2_next_target = qf2_target(data.next_observations, next_state_actions)
-        min_qf_next_target = (
-            torch.min(qf1_next_target, qf2_next_target) - alpha * next_state_log_pi
-        )
-        next_q_value = real_rewards.flatten() + (1 - data.dones.flatten()) * gamma * (
-            min_qf_next_target
-        ).view(-1)
-
-    qf1_a_values = qf1(data.observations, data.actions).view(-1)
-    qf2_a_values = qf2(data.observations, data.actions).view(-1)
-    qf1_loss = F.mse_loss(qf1_a_values, next_q_value)
-    qf2_loss = F.mse_loss(qf2_a_values, next_q_value)
-    qf_loss = qf1_loss + qf2_loss
-
-    # optimize the model
-    q_optimizer.zero_grad()
-    qf_loss.backward()
-    q_optimizer.step()
-
-    return qf_loss, qf1_a_values, qf2_a_values, qf1_loss, qf2_loss
-
-
-def update_actor(data, actor, qf1, qf2, alpha, actor_optimizer):
-    pi, log_pi, _ = actor.get_action(data.observations)
-    qf1_pi = qf1(data.observations, pi)
-    qf2_pi = qf2(data.observations, pi)
-    min_qf_pi = torch.min(qf1_pi, qf2_pi)
-    actor_loss = ((alpha * log_pi) - min_qf_pi).mean()
-
-    actor_optimizer.zero_grad()
-    actor_loss.backward()
-    actor_optimizer.step()
-    if args.autotune:
-        with torch.no_grad():
-            _, log_pi, _ = actor.get_action(data.observations)
-        alpha_loss = (-log_alpha.exp() * (log_pi + target_entropy)).mean()
-
-        a_optimizer.zero_grad()
-        alpha_loss.backward()
-        a_optimizer.step()
-        alpha = log_alpha.exp().item()
-    return actor_loss, alpha, alpha_loss
-
-
-def update_target_networks(source_net, target_net, tau):
-    for param, target_param in zip(source_net.parameters(), target_net.parameters()):
-        target_param.data.copy_(tau * param.data + (1 - tau) * target_param.data)
-
-
-def save_model_all(run_name: str, step: int, state_dict: dict):
-    model_folder = f"./models/{run_name}/{step}"
-    os.makedirs(model_folder, exist_ok=True)
-    out_path = f"{model_folder}/checkpoint.pth"
-
-    total_state_dict = {name: obj.state_dict() for name, obj in state_dict.items()}
-    torch.save(total_state_dict, out_path)
-    logging.info(f"Saved all models and optimizers to {out_path}")
-
-
-def load_model_all(state_dict: dict, path: str, device):
-    assert os.path.exists(path), "Path to model does not exist"
-    checkpoint = torch.load(path, device)
-    for name, obj in state_dict.items():
-        if name in checkpoint:
-            obj.load_state_dict(checkpoint[name])
-    logging.info(f"Models and optimizers loaded from {path}")
-
-
-def save_replay_buffer(run_name: str, step: int, replay_buffer: ReplayBuffer):
-    model_folder = f"./models/{run_name}/{step}"
-    os.makedirs(model_folder, exist_ok=True)
-    out_path = f"{model_folder}/replay_buffer.pth"
-
-    buffer_data = {
-        "observations": replay_buffer.observations,
-        "next_observations": replay_buffer.next_observations,
-        "actions": replay_buffer.actions,
-        "rewards": replay_buffer.rewards,
-        "ground_truth_rewards": replay_buffer.ground_truth_rewards,
-        "dones": replay_buffer.dones,
-    }
-    with gzip.open(out_path, "wb") as f:
-        pickle.dump(buffer_data, f)
-    logging.info(f"Saved replay buffer to {out_path}")
-
-
-def load_replay_buffer(replay_buffer: ReplayBuffer, path: str):
-    assert os.path.exists(path), "Path to replay buffer does not exist"
-    with gzip.open(path, "rb") as f:
-        buffer_data = pickle.load(f)
-
-    replay_buffer.observations = buffer_data["observations"]
-    replay_buffer.next_observations = buffer_data["next_observations"]
-    replay_buffer.actions = buffer_data["actions"]
-    replay_buffer.rewards = buffer_data["rewards"]
-    replay_buffer.ground_truth_rewards = buffer_data["ground_truth_rewards"]
-    replay_buffer.dones = buffer_data["dones"]
-
-    logging.info(f"Replay buffer loaded from {path}")
-
-
-# ALGO LOGIC: initialize agent here:
-class SoftQNetwork(nn.Module):
-    def __init__(self, env, hidden_dim, hidden_layers):
-        super().__init__()
-        self.fc_first = nn.Linear(
-            np.array(env.single_observation_space.shape).prod()
-            + np.prod(env.single_action_space.shape),
-            hidden_dim,
-        )
-
-        self.hidden_layers = nn.ModuleList()
-        for _ in range(hidden_layers):
-            self.hidden_layers.append(nn.Linear(hidden_dim, hidden_dim))
-
-        self.fc_last = nn.Linear(hidden_dim, 1)
-
-    def forward(self, x, a):
-        x = torch.cat([x, a], dim=1)
-        x = F.relu(self.fc_first(x))
-        for layer in self.hidden_layers:
-            x = F.relu(layer(x))
-        x = self.fc_last(x)
-        return x
-
-
-LOG_STD_MAX = 2
-LOG_STD_MIN = -5
-
-
-class Actor(nn.Module):
-    def __init__(self, env, hidden_dim, hidden_layers):
-        super().__init__()
-        self.fc_first = nn.Linear(
-            np.array(env.single_observation_space.shape).prod(), hidden_dim
-        )
-        self.hidden_layers = nn.ModuleList()
-        for _ in range(hidden_layers):
-            self.hidden_layers.append(nn.Linear(hidden_dim, hidden_dim))
-            self.hidden_layers.append(nn.ReLU())
-        self.fc_mean = nn.Linear(hidden_dim, np.prod(env.single_action_space.shape))
-        self.fc_logstd = nn.Linear(hidden_dim, np.prod(env.single_action_space.shape))
-        # action rescaling
-        self.register_buffer(
-            "action_scale",
-            torch.tensor(
-                (env.single_action_space.high - env.single_action_space.low) / 2.0,
-                dtype=torch.float32,
-            ),
-        )
-        self.register_buffer(
-            "action_bias",
-            torch.tensor(
-                (env.single_action_space.high + env.single_action_space.low) / 2.0,
-                dtype=torch.float32,
-            ),
-        )
-
-    def forward(self, x):
-        x = F.relu(self.fc_first(x))
-        for layer in self.hidden_layers:
-            x = layer(x)
-        mean = self.fc_mean(x)
-        log_std = self.fc_logstd(x)
-        log_std = torch.tanh(log_std)
-        log_std = LOG_STD_MIN + 0.5 * (LOG_STD_MAX - LOG_STD_MIN) * (
-            log_std + 1
-        )  # From SpinUp / Denis Yarats
-
-        return mean, log_std
-
-    def get_action(self, x):
-        mean, log_std = self(x)
-        std = log_std.exp()
-        normal = torch.distributions.Normal(mean, std)
-        x_t = normal.rsample()  # for reparameterization trick (mean + std * N(0,1))
-        y_t = torch.tanh(x_t)
-        action = y_t * self.action_scale + self.action_bias
-        log_prob = normal.log_prob(x_t)
-        # Enforcing Action Bound
-        log_prob -= torch.log(self.action_scale * (1 - y_t.pow(2)) + 1e-6)
-        log_prob = log_prob.sum(1, keepdim=True)
-        mean = torch.tanh(mean) * self.action_scale + self.action_bias
-        return action, log_prob, mean
-
-
-if __name__ == "__main__":
+def train(args: Any):
     import stable_baselines3 as sb3
 
     if sb3.__version__ < "2.0":
@@ -371,7 +157,6 @@ poetry run pip install "stable_baselines3==2.0.0a1"
 """
         )
 
-    args = tyro.cli(Args)
     run_name = f"{args.env_id}__{args.exp_name}__{args.seed}__{int(time.time())}"
     logging.basicConfig(
         format="%(asctime)s %(levelname)s: %(message)s",
@@ -413,10 +198,7 @@ poetry run pip install "stable_baselines3==2.0.0a1"
 
     # env setup
     envs = gym.vector.SyncVectorEnv(
-        [
-            make_env(args.env_id, args.seed, i, args.capture_video, run_name)
-            for i in range(args.num_envs)
-        ]
+        [make_env(args.env_id, args.seed) for i in range(args.num_envs)]
     )
     assert isinstance(
         envs.single_action_space, gym.spaces.Box
@@ -553,6 +335,7 @@ poetry run pip install "stable_baselines3==2.0.0a1"
                     alpha,
                     args.gamma,
                     q_optimizer,
+                    actor,
                 )
 
                 if (
@@ -561,8 +344,17 @@ poetry run pip install "stable_baselines3==2.0.0a1"
                     for _ in range(
                         args.policy_frequency
                     ):  # compensate for the delay by doing 'actor_update_interval' instead of 1
-                        actor_loss = update_actor(
-                            data, actor, qf1, qf2, alpha, actor_optimizer
+                        actor_loss, alpha, alpha_loss = update_actor(
+                            data,
+                            actor,
+                            qf1,
+                            qf2,
+                            alpha,
+                            actor_optimizer,
+                            args.autotune,
+                            log_alpha,
+                            target_entropy,
+                            a_optimizer,
                         )
 
             if explore_step % args.target_network_frequency == 0:
@@ -775,6 +567,7 @@ poetry run pip install "stable_baselines3==2.0.0a1"
                     alpha,
                     args.gamma,
                     q_optimizer,
+                    actor,
                 )
 
                 if (
@@ -784,7 +577,16 @@ poetry run pip install "stable_baselines3==2.0.0a1"
                         args.policy_frequency
                     ):  # compensate for the delay by doing 'actor_update_interval' instead of 1
                         actor_loss, alpha, alpha_loss = update_actor(
-                            data, actor, qf1, qf2, alpha, actor_optimizer
+                            data,
+                            actor,
+                            qf1,
+                            qf2,
+                            alpha,
+                            actor_optimizer,
+                            args.autotune,
+                            log_alpha,
+                            target_entropy,
+                            a_optimizer,
                         )
 
                 # update the target networks
@@ -847,3 +649,8 @@ poetry run pip install "stable_baselines3==2.0.0a1"
         save_replay_buffer(run_name, current_step, rb)
         envs.close()
         writer.close()
+
+
+if __name__ == "__main__":
+    cli_args = tyro.cli(Args)
+    train(cli_args)
